# ddtrace/internal.GetGlobalTracer(): Use atomic.Value (91% less CPU)