The GetGlobalTracer() function is called in many places. It currently
uses sync.RWMutex to ensure the current value is read in a
thread-safe fashion. We can replace this with an atomic.Value, which
is substantially more efficient. According to one profile I have, a
program that makes a lot of use of tracing spends approximately 0.5%
of all CPU in this function, so this should be a very small
improvement across all programs.

I tested with two benchmarks: A serial benchmark, which calls
GetGlobalTracer using a single goroutine, and a parallel benchmark,
which calls it from all CPUs at the same time. On my 8 core Macbook,
I get the following results, which show this uses about 91% less CPU
in the serial, uncontended case, or 98% less CPU in the contended,
parallel case. I think the serial, uncontended results are more
representative.
```
name                       old time/op    new time/op    delta
GetGlobalTracerSerial-8      13.2ns ± 2%     1.1ns ± 5%  -91.59%  (p=0.000 n=10+10)
GetGlobalTracerParallel-8    31.9ns ±15%     0.3ns ± 5%  -98.91%  (p=0.000 n=10+10)

name                       old alloc/op   new alloc/op   delta
GetGlobalTracerSerial-8       0.00B          0.00B          ~     (all equal)
GetGlobalTracerParallel-8     0.00B          0.00B          ~     (all equal)

name                       old allocs/op  new allocs/op  delta
GetGlobalTracerSerial-8        0.00           0.00          ~     (all equal)
GetGlobalTracerParallel-8      0.00           0.00          ~     (all equal)
```