@nsrip-dd 2022-04-20
Nice optimization! This change unfortunately requires Go 1.17 for `atomic.(*Value).Swap`. Even if we adopt our [new Go version support policy](https://github.com/DataDog/dd-trace-go/pull/1239), we're still likely going to support Go 1.16 until Go 1.19 comes out. There are a few ways I think we could go here:

* We could still use this change, but have two implementations of the global tracer stuff in different files that are tagged with `//go:build go1.17` and `//go:build !go1.17`. Then people using recent Go releases get this better implementation, and  we can drop the old implementation once we drop support for older Go versions.
* Use the older `atomic.SwapPointer` & friends. This wouldn't be quite as nice as the `atomic.Value` implementation. I don't think there's anything wrong with that in principle but maybe users might squirm at the `unsafe` requirement?

Going the second way could look like this:

```go
var globalTracer unsafe.Pointer

func init() {
	var tracer ddtrace.Tracer = &NoopTracer{}
	atomic.StorePointer(&globalTracer, unsafe.Pointer(&tracer))
}

func SetGlobalTracer(t ddtrace.Tracer) {
	old := (*ddtrace.Tracer)(atomic.SwapPointer(&globalTracer, unsafe.Pointer(&t)))
	if !Testing {
		(*old).Stop()
	}
}

func GetGlobalTracer() ddtrace.Tracer {
	return *(*ddtrace.Tracer)(atomic.LoadPointer(&globalTracer))
}
```

Any thoughts on what's better?