@evanj 2022-06-06
Sorry for the delay, I was on parental leave. Thanks for the careful reviews! I did not notice that `atomic.Value.Swap` is recent. At this point, maybe another answer is to wait until August when Go 1.19 is released, assuming the support policy will drop Go 1.16 then?

I like your SwapPointer solution also. I think it should be equivalent to the `Value.Swap` solution? Either of those solutions seem good to me.